@@ -16,6 +16,8 @@ import { Media } from '@ionic-native/media/ngx';
 import { File } from '@ionic-native/File/ngx';
 import { CallNumber } from '@ionic-native/call-number/ngx';
 // import { CallLog } from '@ionic-native/call-log/ngx';
+ import { DocumentViewer } from '@ionic-native/document-viewer/ngx';
+ import { FileTransfer } from '@ionic-native/file-transfer/ngx';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DatePipe, CurrencyPipe} from '@angular/common';
 import { FileOpener } from '@ionic-native/file-opener/ngx';
@@ -24,6 +26,10 @@ import { NativeGeocoder, NativeGeocoderResult, NativeGeocoderOptions } from '@io
 import { AuthInterceptor } from './shared/auth.interceptor';
 import { FocusDirective } from './directive/focus.directive';
 import { IonicCoreModule } from '@core/ionic-core';
+
+ 
+
+
 
 
 
@@ -34,14 +40,16 @@ import { IonicCoreModule } from '@core/ionic-core';
     BrowserModule, 
     FormsModule, 
     ReactiveFormsModule, 
-    HttpClientModule, 
-    IonicModule.forRoot(), 
+    HttpClientModule,
+    IonicModule.forRoot(),
     IonicCoreModule.forRoot(environment),
-    AppRoutingModule, 
+    AppRoutingModule,
     ServiceWorkerModule.register('ngsw-worker.js', 
     { enabled: environment.production })
   ],
   providers: [
+    DocumentViewer,
+    FileTransfer,
     StatusBar,
     SplashScreen,
     MediaCapture,
