@@ -8,21 +8,21 @@ import { Router } from '@angular/router';
 export class ContactPage implements OnInit {
   modal: any;
   userdata = [
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
-    // {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'}
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'},
+    {'name': 'Amit Tomer', 'profile': 'Project Manager', 'company': 'E-Labs', 'mobile': '9205280930'}
   ]
  
   constructor(private router: Router,) { }
