@@ -11,10 +11,20 @@ import { Subscription } from 'rxjs';
 import * as appConstants from '../../app//shared/app.constants';
 import { ProductSearchComponent } from '../component/product-search/product-search.component';
 
+import { DocumentViewer, DocumentViewerOptions } from '@ionic-native/document-viewer/ngx';
+import { File } from '@ionic-native/file/ngx';
+import { FileTransfer } from '@ionic-native/file-transfer/ngx';
+
+import { FileOpener } from '@ionic-native/file-opener/ngx';
+import { Plugins } from '@capacitor/core';
+import { FilesystemDirectory } from '@capacitor/filesystem';
+
+
 @Component({
   selector: 'app-home',
   templateUrl: './home.page.html',
   styleUrls: ['./home.page.scss'],
+  providers: [File]
 })
 export class HomePage implements OnInit {
   modal: any;
@@ -48,7 +58,11 @@ export class HomePage implements OnInit {
 
 
   @Output() collection_name = new EventEmitter<string>();
+  plt: any;
+  pdfObj: any;
 
+
+  // download var
 
   constructor(
     private platform: Platform,
@@ -63,7 +77,11 @@ export class HomePage implements OnInit {
     private loaderService: LoaderService,
     private envService: EnvService,
     private permissionService: PermissionService,
-    private dataShareService: DataShareServiceService
+    private dataShareService: DataShareServiceService,
+    private fileOpener: FileOpener,
+    private documentViewer: DocumentViewer,
+    private file: File,
+    private fileTransfer: FileTransfer
   ) 
   {
     // below code is for slider and title name
@@ -316,5 +334,40 @@ export class HomePage implements OnInit {
   comingSoon() {
     this.storageService.presentToast('Comming Soon...');
   }
+
+
+  pdfurl = "https://file-examples-com.github.io/uploads/2017/10/file-sample_150kB.pdf";
+ 
+  downloadPDF(){
+    const options: DocumentViewerOptions = {
+      title: 'My PDF'
+    }
+    this.documentViewer.viewDocument('https://file-examples-com.github.io/uploads/2017/10/file-sample_150kB.pdf', 'application/pdf', options);
+    console.log(this.pdfurl);
+  }
+
+  openPDF(){
+    const options: DocumentViewerOptions = {
+      title: 'My PDF'
+    }
+    this.documentViewer.viewDocument('https://file-examples-com.github.io/uploads/2017/10/file-sample_150kB.pdf', 'aplication/pdf', options )
+  }
+  downloadAndopenPDF(){
+    let path = null;
+
+    if(this.platform.is('ios')){
+      path = this.file.documentsDirectory;
+    }else{
+      path = this.file.dataDirectory;
+    }
+
+    const fileTransfer = this.fileTransfer.create();
+    fileTransfer.download('https://file-examples-com.github.io/uploads/2017/10/file-sample_150kB.pdf', path + 'myfile.pdf').then(entry => {
+    let url = entry.toURL();
+    this.documentViewer.viewDocument(url, 'application/pdf', {});
+
+    })
+
+  }  
 
 }
