@@ -5,8 +5,8 @@
 export const environment = {
   production: false,
   baseUrl: 'https://omoknow.com/rest/',
-  serverhost: "https://uatserverqualiteklab.e-labs.ai/rest/",
-  // serverhost: "http://localhost:8104/rest/",
+  //serverhost: "https://uatserverqualiteklab.e-labs.ai/rest/",
+   serverhost: "http://localhost:8104/rest/",
   mediceaHost: "http://pcp.medicea.in/rest/",
   appId: "ELABS",
   appName: "tonicprod",
